Add files via upload
We made the source code of this prediction model publicly available so that more researchers could refer to and use it. This code includes data preprocessing, model building, performance evaluation, and implementation of SHAP interpretation methods. Interested researchers are encouraged to review and use these codes to learn how to build and optimize predictive models on similar medical datasets.